+++ src/components/Card.js
@@ -3,11 +3,8 @@ import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import Box from "@material-ui/core/Box";
-// import useStyles from "./styles/useStyles";
-// import ButtonBase from "@material-ui/core/ButtonBase";
 
 const Card = (props) => {
-  // const classes = useStyles();
   const { releaseDate, releaseInfo } = props;
   const infoPapers = releaseInfo[releaseDate].map((item, index) => (
     <Paper key={index}>
